testing/mochitest/baselinecoverage/plain/test_baselinecoverage.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Baseline Coverage Collection</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type="application/javascript">
  SimpleTest.ok(true, "Collecting  baseline coverage for HTML (.html) file types.");
</script>
</body>
</html>
